Refactor committee page UI and add new team photo
This commit refactors the UI of the committee page to make it more user-friendly. It also adds a new team photo to the WebDevelopmentTeam component. The width of the lead member's picture is adjusted to 80% for better visibility. The previous team photo is replaced with a new one named teamPhoto2.